fix: return error response instead of raising internal server error

+++ insta_milligram/users/views/delete_view.py
@@ -5,7 +5,12 @@ import rest_framework.response as rr  # type: ignore
 import rest_framework.status as rs  # type: ignore
 
 
-def delete(request: dr.HttpRequest, id: int):
+def delete(request: dr.HttpRequest, id: int = -1):
+    if id == -1:
+        return rr.Response(
+            {"message": "User ID Missing"},
+            rs.HTTP_400_BAD_REQUEST,
+        )
     try:
         dam.User.objects.get(pk=id).delete()
         return rr.Response({"message": "Success"}, rs.HTTP_200_OK)
